refactor bookings functionality and update API endpoints; remove deprecated bookings components
